Move StateInternalsFactory to runners/core-java
