+++ runners/core-java/src/main/java/org/apache/beam/runners/core/StateInternalsFactory.java
@@ -15,11 +15,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util.state;
+package org.apache.beam.runners.core;
 
 import java.io.Serializable;
 import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.annotations.Experimental.Kind;
+import org.apache.beam.sdk.util.state.StateInternals;
 
 /**
  * A factory for providing {@link StateInternals} for a particular key.
